More api authentication
