+++ src/pages/api/admin/[id].ts
@@ -3,7 +3,7 @@ import { HTTP_STATUS_CODE } from "@/utils/consts";
 import { AdminInvalidInputException } from "@/utils/exceptions/admin";
 import mongoose from "mongoose";
 import { NextApiRequest, NextApiResponse } from "next";
-import { authenticate } from "@/middleware";
+import { authenticate } from "../auth/[...nextauth]";
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse,
