+++ src/pages/api/auth/[...nextauth].ts
@@ -2,7 +2,10 @@ import NextAuth, { NextAuthOptions } from "next-auth";
 import { verifyUser } from "@/server/db/actions/UserAction";
 import CredentialsProvider from "next-auth/providers/credentials";
 import { loginSchema } from "@/components/Login/LoginForm";
+import { NextApiRequest, NextApiResponse } from "next";
 
+import { getServerSession } from "next-auth";
+import { HTTP_STATUS_CODE } from "@/utils/consts";
 /**
  *  @type {import("next-auth").NextAuthOptions}
  */
@@ -69,3 +72,27 @@ export const authOptions: NextAuthOptions = {
 };
 
 export default NextAuth(authOptions);
+
+export const authenticate = async (
+  req: NextApiRequest,
+  res: NextApiResponse,
+  methods: string[],
+  adminRequired: boolean = false,
+) => {
+  if (req.method == undefined) {
+    return res.status(HTTP_STATUS_CODE.METHOD_NOT_ALLOWED).json({
+      error: `No request method provided`,
+    });
+  }
+  if (methods.includes(req.method)) {
+    const session = await getServerSession(req, res, authOptions);
+    if (session && (session.user.label == "administator" || !adminRequired)) {
+      return true;
+    } else {
+      return res.status(HTTP_STATUS_CODE.UNAUTHORIZED).json({
+        error: "Unauthorized",
+      });
+    }
+  }
+  return true;
+};
