+++ src/pages/api/tags/index.ts
@@ -8,7 +8,7 @@ import {
 import { HTTP_STATUS_CODE } from "@/utils/consts";
 import { z } from "zod";
 import { TagException, TagInvalidInputException } from "@/utils/exceptions/tag";
-import { authenticate } from "@/middleware";
+import { authenticate } from "../auth/[...nextauth]";
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse,
